+++ source/game/g_syscalls.h
@@ -68,7 +68,7 @@ static inline int trap_SkinIndex( const char *name ) {
 	return GAME_IMPORT.SkinIndex( name );
 }
 
-static inline unsigned int trap_Milliseconds( void ) {
+static inline int64_t trap_Milliseconds( void ) {
 	return GAME_IMPORT.Milliseconds();
 }
